deps: Intl: ICU 58 bump - small icu (BIG COMMIT)
This commit contains the ICU 58.1 delta.
It is especially large because of the ICU license change,
and, because the line endings were off previously.

* bump to ICU 58.1 - check in small ICU source
* from 58.1 final http://site.icu-project.org/download/58

Fixes: https://github.com/nodejs/node/issues/7844
PR-URL: https://github.com/nodejs/node/pull/9234

+++ deps/icu-small/source/i18n/rbt_rule.cpp
@@ -1,3 +1,5 @@
+// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
  **********************************************************************
  *   Copyright (C) 1999-2011, International Business Machines
@@ -124,7 +126,7 @@ TransliterationRule::TransliterationRule(const UnicodeString& input,
             return;
         }
     }
-
+    
     key = NULL;
     if (keyLength > 0) {
         key = new StringMatcher(pattern, anteContextLength, anteContextLength + keyLength,
@@ -135,7 +137,7 @@ TransliterationRule::TransliterationRule(const UnicodeString& input,
             return;
         }
     }
-
+    
     int32_t postContextLength = pattern.length() - keyLength - anteContextLength;
     postContext = NULL;
     if (postContextLength > 0) {
@@ -174,7 +176,7 @@ TransliterationRule::TransliterationRule(TransliterationRule& other) :
     segmentsCount = 0;
     if (other.segmentsCount > 0) {
         segments = (UnicodeFunctor **)uprv_malloc(other.segmentsCount * sizeof(UnicodeFunctor *));
-        uprv_memcpy(segments, other.segments, other.segmentsCount*sizeof(segments[0]));
+        uprv_memcpy(segments, other.segments, (size_t)other.segmentsCount*sizeof(segments[0]));
     }
 
     if (other.anteContext != NULL) {
@@ -257,7 +259,7 @@ UBool TransliterationRule::masks(const TransliterationRule& r2) const {
      * r1:      aakkkpppp
      * r2:     aaakkkkkpppp
      *            ^
-     *
+     * 
      * The strings must be aligned at the first character of the
      * key.  The length of r1 to the left of the alignment point
      * must be <= the length of r2 to the left; ditto for the
@@ -335,10 +337,10 @@ static inline int32_t posAfter(const Replaceable& str, int32_t pos) {
  * context and key characters match, but the text is not long
  * enough to match all of them.  A full match means all context
  * and key characters match.
- *
+ * 
  * If a full match is obtained, perform a replacement, update pos,
  * and return U_MATCH.  Otherwise both text and pos are unchanged.
- *
+ * 
  * @param text the text
  * @param pos the position indices
  * @param incremental if TRUE, test for partial matches that may
@@ -381,7 +383,7 @@ UMatchDegree TransliterationRule::matchAndReplace(Replaceable& text,
     // 32-bit code points.  This works because stand-ins are
     // always in the BMP and because we are doing a literal match
     // operation, which can be done 16-bits at a time.
-
+    
     int32_t anteLimit = posBefore(text, pos.contextStart);
 
     UMatchDegree match;
@@ -399,13 +401,13 @@ UMatchDegree TransliterationRule::matchAndReplace(Replaceable& text,
     minOText = posAfter(text, oText);
 
     // ------------------------ Start Anchor ------------------------
-
+    
     if (((flags & ANCHOR_START) != 0) && oText != anteLimit) {
         return U_MISMATCH;
     }
 
     // -------------------- Key and Post Context --------------------
-
+    
     oText = pos.start;
 
     if (key != NULL) {
@@ -431,9 +433,9 @@ UMatchDegree TransliterationRule::matchAndReplace(Replaceable& text,
             return match;
         }
     }
-
+    
     // ------------------------- Stop Anchor ------------------------
-
+    
     if (((flags & ANCHOR_END)) != 0) {
         if (oText != pos.contextLimit) {
             return U_MISMATCH;
@@ -442,7 +444,7 @@ UMatchDegree TransliterationRule::matchAndReplace(Replaceable& text,
             return U_PARTIAL_MATCH;
         }
     }
-
+    
     // =========================== REPLACE ==========================
 
     // We have a full match.  The key is between pos.start and
