+++ deps/icu-small/source/i18n/ucoleitr.cpp
@@ -1,3 +1,5 @@
+// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
 ******************************************************************************
 *   Copyright (C) 2001-2016, International Business Machines
@@ -9,7 +11,7 @@
 * Modification History:
 *
 * Date        Name        Description
-* 02/15/2001  synwee      Modified all methods to process its own function
+* 02/15/2001  synwee      Modified all methods to process its own function 
 *                         instead of calling the equivalent c++ api (coleitr.h)
 * 2012-2014   markus      Rewritten in C++ again.
 ******************************************************************************/
@@ -34,9 +36,9 @@ U_NAMESPACE_USE
 #define DEFAULT_BUFFER_SIZE 16
 #define BUFFER_GROW 8
 
-#define ARRAY_COPY(dst, src, count) uprv_memcpy((void *) (dst), (void *) (src), (count) * sizeof (src)[0])
+#define ARRAY_COPY(dst, src, count) uprv_memcpy((void *) (dst), (void *) (src), (size_t)(count) * sizeof (src)[0])
 
-#define NEW_ARRAY(type, count) (type *) uprv_malloc((count) * sizeof(type))
+#define NEW_ARRAY(type, count) (type *) uprv_malloc((size_t)(count) * sizeof(type))
 
 #define DELETE_ARRAY(array) uprv_free((void *) (array))
 
@@ -311,7 +313,7 @@ ucol_reset(UCollationElements *elems)
 }
 
 U_CAPI int32_t U_EXPORT2
-ucol_next(UCollationElements *elems,
+ucol_next(UCollationElements *elems, 
           UErrorCode         *status)
 {
     if (U_FAILURE(*status)) {
@@ -395,7 +397,7 @@ UCollationPCE::previousProcessed(
         // buffer raw CEs up to non-ignorable primary
         RCEBuffer rceb;
         int32_t ce;
-
+        
         // **** do we need to reset rceb, or will it always be empty at this point ****
         do {
             high = cei->getOffset();
@@ -431,14 +433,14 @@ UCollationPCE::previousProcessed(
 finish:
     if (pceBuffer.isEmpty()) {
         // **** Is -1 the right value for ixLow, ixHigh? ****
-	if (ixLow != NULL) {
-		*ixLow = -1;
-	}
-
-	if (ixHigh != NULL) {
-		*ixHigh = -1
-		;
-	}
+    	if (ixLow != NULL) {
+    		*ixLow = -1;
+    	}
+    	
+    	if (ixHigh != NULL) {
+    		*ixHigh = -1
+    		;
+    	}
         return UCOL_PROCESSED_NULLORDER;
     }
 
@@ -509,19 +511,19 @@ ucol_setOffset(UCollationElements    *elems,
 }
 
 U_CAPI int32_t U_EXPORT2
-ucol_primaryOrder (int32_t order)
+ucol_primaryOrder (int32_t order) 
 {
     return (order >> 16) & 0xffff;
 }
 
 U_CAPI int32_t U_EXPORT2
-ucol_secondaryOrder (int32_t order)
+ucol_secondaryOrder (int32_t order) 
 {
     return (order >> 8) & 0xff;
 }
 
 U_CAPI int32_t U_EXPORT2
-ucol_tertiaryOrder (int32_t order)
+ucol_tertiaryOrder (int32_t order) 
 {
     return order & 0xff;
 }
