+++ deps/icu-small/source/io/sprintf.c
@@ -1,3 +1,5 @@
+// Copyright (C) 2016 and later: Unicode, Inc. and others.
+// License & terms of use: http://www.unicode.org/copyright.html
 /*
 ******************************************************************************
 *
@@ -62,7 +64,7 @@ u_sprintf_pad_and_justify(void                        *context,
     if(info->fWidth != -1 && resultLen < info->fWidth) {
         int32_t paddingLeft = info->fWidth - resultLen;
         int32_t outputPos = output->len - output->available;
-
+  
         if (paddingLeft + resultLen > output->available) {
             paddingLeft = output->available - resultLen;
             if (paddingLeft < 0) {
@@ -89,9 +91,9 @@ u_sprintf_pad_and_justify(void                        *context,
     else {
         written = u_sprintf_write(output, result, resultLen);
     }
-
+    
     if (written >= 0 && lengthOfResult > written) {
-	return lengthOfResult;
+    	return lengthOfResult;
     }
 
     return written;
@@ -179,7 +181,7 @@ u_vsnprintf(UChar       *buffer,
     int32_t size = (int32_t)strlen(patternSpecification) + 1;
 
     /* convert from the default codepage to Unicode */
-    if (size >= MAX_UCHAR_BUFFER_SIZE(patBuffer)) {
+    if (size >= (int32_t)MAX_UCHAR_BUFFER_SIZE(patBuffer)) {
         pattern = (UChar *)uprv_malloc(size * sizeof(UChar));
         if(pattern == 0) {
             return 0;
@@ -201,13 +203,13 @@ u_vsnprintf(UChar       *buffer,
     return written;
 }
 
-U_CAPI int32_t U_EXPORT2
-u_vsprintf_u(UChar       *buffer,
-             const UChar *patternSpecification,
-             va_list     ap)
-{
-    return u_vsnprintf_u(buffer, INT32_MAX, patternSpecification, ap);
-}
+U_CAPI int32_t U_EXPORT2 
+u_vsprintf_u(UChar       *buffer, 
+             const UChar *patternSpecification, 
+             va_list     ap) 
+{ 
+    return u_vsnprintf_u(buffer, INT32_MAX, patternSpecification, ap); 
+} 
 
 static const u_printf_stream_handler g_sprintf_stream_handler = {
     u_sprintf_write,
@@ -239,7 +241,7 @@ u_vsnprintf_u(UChar    *buffer,
 
     /* parse and print the whole format string */
     result = u_printf_parse(&g_sprintf_stream_handler, patternSpecification, &outStr, &outStr, &outStr.fBundle, &written, ap);
-
+    
     /* Terminate the buffer, if there's room. */
     if (outStr.available > 0) {
         buffer[outStr.len - outStr.available] = 0x0000;
@@ -248,12 +250,13 @@ u_vsnprintf_u(UChar    *buffer,
     /* Release the cloned bundle, if we cloned it. */
     u_locbund_close(&outStr.fBundle);
 
-    /* parsing error */
+    /* parsing error */ 
     if (result < 0) {
-	return result;
+    	return result;
     }
     /* return # of UChars written */
     return written;
 }
 
 #endif /* #if !UCONFIG_NO_FORMATTING */
+
